+++ packages/enhanced/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @module-federation/enhanced
 
+## 0.17.0
+
+### Patch Changes
+
+- Updated dependencies [3f736b6]
+- Updated dependencies [e0ceca6]
+- Updated dependencies [1d691ef]
+- Updated dependencies [60e4df1]
+  - @module-federation/data-prefetch@0.17.0
+  - @module-federation/cli@0.17.0
+  - @module-federation/dts-plugin@0.17.0
+  - @module-federation/runtime-tools@0.17.0
+  - @module-federation/manifest@0.17.0
+  - @module-federation/rspack@0.17.0
+  - @module-federation/inject-external-runtime-core-plugin@0.17.0
+  - @module-federation/sdk@0.17.0
+  - @module-federation/managers@0.17.0
+  - @module-federation/bridge-react-webpack-plugin@0.17.0
+  - @module-federation/error-codes@0.17.0
+
 ## 0.16.0
 
 ### Patch Changes